+++ src/components/Common/PaperListEmojiBadge/PaperListEmojiBadge.jsx
