+++ src/components/Common/PaperListEmojiBadge/index.js
